requirements.txt
openai
langchain
python-dotenv
slack_bolt